enlarge dev_t to 32-bit on i386, arm, sparc, s390 (Pascal Terjan)   update sys/stat.h to include nsec and enlarge dev_t

+++ include/sys/types.h
@@ -45,7 +45,7 @@ typedef uint64_t fsfilcnt_t;
     typedef uint32_t nlink_t;		/* Used for link counts. */
     typedef uint32_t uid_t;		/* Used for user IDs. */
 #elif defined(__arm__) || defined(__i386__) || defined(__sparc__) || defined(__s390__) /* make sure __s390x__ hits before __s390__ */
-    typedef uint16_t dev_t;
+    typedef uint32_t dev_t;
     typedef uint16_t gid_t;
     typedef uint16_t mode_t;
     typedef uint16_t nlink_t;
